Moving More Core Attributes
