+++ src/Views/Traits/IsFilter.php
@@ -4,7 +4,7 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits;
 
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\FilterConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\FilterHelpers;
-use Rappasoft\LaravelLivewireTables\Views\Traits\{HasConfig,HasView};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasConfig,HasView};
 
 trait IsFilter
 {
